Refactor to put the event handling stuff into EventUtils
